feat: move checkForUpdate to pkg/cmd/version/options.go
Refs: #408
