+++ pkg/version/types_test.go
@@ -1,3 +1,6 @@
+//go:build !arm64
+// +build !arm64
+
 package version
 
 import (
